Update README.md
Programming languages such as Python and R offer flexibility in the choice of data analysis software. Unfortunately, many faculties in public higher education institutions do not provide adequate training opportunities for learning these languages, which can and should be important tools for the exploitation of scientific data. To meet their needs, some teachers and researchers turn to paid software. Unfortunately, due to limited resources, they often make do with old versions, most of which were acquired through projects and are now more than 20 years old. The lack of updates, due to financial constraints, sometimes makes the use of these software programs unsuitable for current training and research needs.

The present application, currently under development, aims to address this need. It will enable the entire scientific community to have access to a tool that meets their requirements for data exploitation and time efficiency. It is being developed using the Shiny framework of the R language.

One of the objectives of this application is to provide formatted and structured results for direct use.